rename KnowledgePanelsMap -> KnowledgePanels and formmating changes
