+++ test/knowledge_panels_from_json_test.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:openfoodfacts/model/KnowledgePanel.dart';
-import 'package:openfoodfacts/model/KnowledgePanelsMap.dart';
+import 'package:openfoodfacts/model/KnowledgePanels.dart';
 import 'package:test/test.dart';
 
 void main() {
